improve documentation for Julia
the buffer *Completions* gets removed after pressing enter and handled
by other calls, e.g. it gets rewritten when tab-completion is used and
not just append the new stuff, so I use this for now.

For future use this should be improved and a own Buffer for
Documentation should be used