fix touchpad
Since the name of the touchpad changes and also the device id is not
guaranteed to be the same, e.g. when another mouse is attached during
start of X server, some work has to be done in order to find the right
device. Now some minimal sed script is used.