awesome: add new and improved widgets
The volumearc has been replaced by an improved version without memory
leaks (hopefully).  The api of aweseome is now better used, for
example see the update of graphics when changed by mouse interactions.

For CPU and batteryarc the changes are not so big. Due to the
different interaction possibilities, i.e. showing a pop-up only, there
are not so many possible improvments.

Note that all code for the widgets is heavily inspired by
https://github.com/streetturtle/awesome-wm-widgets from Pavel Makhov.