improve open-pdf
Make better checks for invalid page numbers. If no page number is
given and the document is to be opened open it without giving zathura
a page number. This results in opening the file at the last position,
it integrates better with everything else.

Let pgrep do some more work, this removes the necessity of using grep
together with pgrep. It is not the best solution one could have but it
results in the same Murks as before.