add emacs function open-pdf
This function opens a PDF file at a given page, or if it is already
open it jumps to this page.

It uses D-Bus to jumpt to the page, because zathura does not have
anouther useful interface for this. Also the process id has to be
retrieved from the system, as zathura has this in its names for D-Bus.