emacs: change configuration
Use literate programming in Orgmode for the configuration file. This
implies ~/.emacs file should contain just a few lines, e.g., setup
package management, load require org, and load the org file with all
the rest.

The advantage is clarity in the config.org file. Folding the Org tree
works very well and shows just the necessary parts of the file. Also
the explanations are better to write and read in Org mode than as elisp comments.

Note: some work still has to be done. The .emacs file is not as tidy
as it should be.