Emacs: extend mu4e config
refile-folder is now defined for one context. This moves messages to
an archive corresponding to the year and month the message was sent.

The big code block is splitted into smaller parts and there is more
explanation. This explanation now also goes with the relevant code
parts.
In addition the changed code now has better indent.