Emacs: add mu4e-query-rewrite-function
This function is executed before the query is sent to mu. At this
point just add the maildir oft the current context if nothing like
another maildir or a global search (I introduced this here) was
specified.

This function would clearly benefit from setting a root maildir for
contexts. It will pay off to investigate on this.