Emacs: improve indentation
Use spaces, they are more relyable and portable.

With some old setting for Org mode I had to set the lisp-indent-offset
to a specific value in order to get some slighlty useful result. This
is not a good idea if for doing real work with Lisp. I think the
tab-acts-natively setting changed a lot in Org mode.