Emacs: make yasnippet load my own snippets
Yasnippet does not load all the snippets from the custom snippets
directory by default. These are only loaded if yas-global-mode is
enabled or if yas-reload-all is invoked. (yas-global-mode does this
automatically) As I want to select the major modes where yasnippet is
enabled by hand, I have to reload the snippets manually.