Emacs: improve mu4e refile
mu4e refile used to just remove the new flag (and maybe set a seen
flag) and then moves the message. This leads to many mails flagged as
unread in the archive. When the mail goes to the archive is is to be
treated as read by default, otherwise the searches will blow up.

This was also a nice lesson in why Lisp is such a good language for
dynamic programs. Here I had some list of lists that define the
actions to perform, and I could easily change the needed list entry.