Emacs: extend mail settings
The accompanying text for some of the more usual settings is improved
and extended, so it is sayd what is done and why.

Use imagemagick to display images rather than some other program (up
to now mostly xpdf, which I do not like) choosen by some hidden logic
contrary to all OS settings.

Support bogofilter for classifying mails as spam. Here some functions
are added that hand the mails to bogofilter for registering as spam or
ham. Also a `mu4e-junk-folder` has been added, it is not part of mu4e
although it looks as it would be.