msmtp: add msmtpqueue folder and links to the scripts
These scripts are provided by the package including the msmtp binary. They do
not execute very much work. Basically they store the message files along with
a file containing the flags and switches for msmtp. In the end there is a
command for processing the whole folder and sending every mail file with the
accomanying command.