Emacs: add new enhanced outline mode
This mode adds some funcitonality to outline mode like auto-folding with TAB
key, as known from Org mode. This mode features numbering the Headlines as it is
done in usual text.

For the future there may also be something like a autonumbering, mostly for
inserting new headlines directly without entering the numbers manually. This
would also allow for moving headlines or sections.