Emacs: show more mails as text mails
mu4e does not allow to show the text version of a mail, if the text is less than
some ratio of the HTML version. This ratio has to be tested for the mails you
get, some have a huge HTML part, but just a few lines of text (even in the HTML
version).