Emacs: change the font setup
The developper of the Fira Code Font added some recommendations about the usage
in Emacs. This seems to be a better choice than the old setup.

I had to change the Fira Code minor mode in order to respect my distinction of
graphic mode and TTY use. So the mode does meaningful work only in graphic mode.